Add style index.js files to optionally consume the CSS via a js module import.
The webpack css-loader plugin does not deduplicate css that is imported via the css @import. Making the css available via a js module import lets webpack do deduplication at the module level in a way that respects the dependency ordering.

Adding these index.js files should not affect anything in Lumino itself, but just make it easier to consume the Lumino CSS and have deduplication.

This will help fix a problem in JupyterLab, for example, where the Lumino widget CSS is duplicated on the page dozens of times because many modules require the CSS files directly (so css-loader is used). With this change, modules will be able to require the style/index.js module, so Webpack itself will deduplicate the imports, and the Lumino CSS will only be put on the page once.

The default-theme package needs a more complicated index.js since it uses @import in its index.css (in this case, css-loader would make a copy of the css).