Switch to using `setTimeout` for all deferred processing.
As noted on issue #392, using `requestAnimationFrame` means that
no deferred tasks can take place while the user has the browser window
in the background. This can lead to a bad experience for long-lasting
tasks, particularly setup tasks (when a user might start JupyterLab,
switch somewhere else when it loads, and then come back).

This change instead uses `setTimeout` (with a timeout of zero, either
implicitly or explicitly), which immediately puts the task on the event
loop to be run (subject to less severe browser throttling). This allows
these tasks to occur in the background (and should hopefully not be too
taxing on user resources).

https://developer.mozilla.org/en-US/docs/Web/API/setTimeout#timeouts_in_inactive_tabs

> Firefox Desktop and Chrome both have a minimum timeout of 1 second
> for inactive tabs.