Use Actual Return character for Mac shortcuts.
The current character is "Leftwards Arrow with Hook" U+21A9
https://www.compart.com/en/unicode/U+21A9  from the "Arrows" block
which often is rendered as colored emoji (CF code where U+FE0F Variation
Selector-16) was added.

Replace with "Return Symbol" U+23CE from the "Miscellaneous Technical",
which is more consistent as many of the other characters in this list
are also from the "Miscellaneous Technical" section (Backspace, Alt,
Shift..)

This should not need the Variation Selector and will have a similar look
to Shift/Maj... in most fonts.